Rename module and package to com.lithiumsheep.weatherwrapper
