+++ app/build.gradle
@@ -38,7 +38,7 @@ dependencies {
     })
     testCompile 'junit:junit:4.12'
 
-    compile project(":weatherwrapperwhuut")
+    compile project(':weatherwrapper')
 
     // https://www.reddit.com/r/androiddev/comments/6fww6h/play_services_11_released/
     //compile 'com.google.android.gms:play-services-location:11.0.2'
