+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':weatherwrapperwhuut'
+include ':app', ':weatherwrapper'
