+++ weatherwrapper/src/main/java/com/lithiumsheep/weatherwrapper/api/WeatherApiManager.java
@@ -1,4 +1,4 @@
-package com.lithiumsheep.weatherwrapperwhuut.api;
+package com.lithiumsheep.weatherwrapper.api;
 
 import android.location.Location;
 import android.support.annotation.RestrictTo;
