+++ weatherwrapper/src/main/res/values/strings.xml
